+++ test/sources/new-0.8.20.sol
@@ -1,0 +1,23 @@
+pragma solidity >=0.8.20;
+
+contract T {
+    event E();
+
+    function test() external {
+        // ContractDefinition.usedEvents
+        emit E();
+
+        // ContractDefinition.internalFunctionIDs
+        (foo)();
+    }
+
+    function foo() internal {
+    }
+
+
+    // StructDefinition.documentation
+    /// docs
+    struct A {
+        uint x;
+    }
+}