+++ test/sources/new-0.8.21.sol
@@ -1,0 +1,3 @@
+pragma solidity >=0.8.21;
+
+pragma experimental solidity;