+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Significantly optimized `findAll` and `astDereferencer`.
 - Added ability to enumerate all nodes with `findAll('*', node)` or dereference an unknown node type with `deref('*', id)`.
 - Fixed `findAll` exhaustivity around objects in the AST that are not AST nodes.
+- Added `SourceUnit.experimentalSolidity`, available since Solidity 0.8.21.
+- Added `ContractDefinition.usedEvents` and `ContractDefinition.internalFunctionIDs`, available since Solidity 0.8.20.
+- Added `StructDefinition.documentation`, available since Solidity 0.8.20.
+- Added `'shanghai'` as a possible value for `InlineAssembly.evmVersion`.
+- Added `nativeSrc` to Yul AST nodes.
 
 ### 0.4.49 (2023-05-04)